+++ packages/espressocash_app/lib/features/outgoing_split_key_payments/src/bl/oskp_service.dart
@@ -8,6 +8,7 @@ import 'package:uuid/uuid.dart';
 import '../../../../config.dart';
 import '../../../../core/accounts/bl/ec_wallet.dart';
 import '../../../../core/amount.dart';
+import '../../../../core/escrow_private_key.dart';
 import '../../../../core/transactions/resign_tx.dart';
 import '../../../../core/transactions/tx_sender.dart';
 import '../../models/outgoing_split_key_payment.dart';
