+++ packages/espressocash_app/lib/features/outgoing_split_key_payments/src/bl/repository.dart
@@ -10,6 +10,7 @@ import 'package:solana/encoder.dart';
 
 import '../../../../core/amount.dart';
 import '../../../../core/currency.dart';
+import '../../../../core/escrow_private_key.dart';
 import '../../../../core/tokens/token_list.dart';
 import '../../../../core/transactions/tx_sender.dart';
 import '../../../../data/db/db.dart';
